+++ PsychMonkey7/PsychMonkey7-webapp-exec/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>PsychMonkey7</name>
+	<name>PsychMonkey7-webapp-exec</name>
 	<comment></comment>
 	<projects>
 	</projects>
