+++ PsychMonkey7/PsychMonkey7-webapp-it/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>PsychMonkey7</name>
+	<name>PsychMonkey7-webapp-it</name>
 	<comment></comment>
 	<projects>
 	</projects>
