--- PsychMonkey7/PsychMonkey7a-api-impl/src/main/resources/META-INF/spring-context.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0"?>
-
-<!--
-  ~ Licensed to the Apache Software Foundation (ASF) under one
-  ~ or more contributor license agreements.  See the NOTICE file
-  ~ distributed with this work for additional information
-  ~ regarding copyright ownership.  The ASF licenses this file
-  ~ to you under the Apache License, Version 2.0 (the
-  ~ "License"); you may not use this file except in compliance
-  ~ with the License.  You may obtain a copy of the License at
-  ~
-  ~   http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing,
-  ~ software distributed under the License is distributed on an
-  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~ KIND, either express or implied.  See the License for the
-  ~ specific language governing permissions and limitations
-  ~ under the License.
-  -->
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:context="http://www.springframework.org/schema/context"
-       xmlns:jaxrs="http://cxf.apache.org/jaxrs"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-           http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-           http://www.springframework.org/schema/context 
-           http://www.springframework.org/schema/context/spring-context-3.0.xsd
-           http://cxf.apache.org/jaxrs
-           http://cxf.apache.org/schemas/jaxrs.xsd">
-
-  <import resource="classpath:META-INF/cxf/cxf.xml"/>
-  <!--
-  <import resource="classpath:META-INF/cxf/cxf-extension-jaxrs-binding.xml"/>
-  -->
-  <import resource="classpath:META-INF/cxf/cxf-servlet.xml"/>
-
-  <context:annotation-config />
-  <context:component-scan base-package="edu.asu.bmi.cad.PsychMonkey7a"/>
-
-  <jaxrs:server id="testServices" address="/testServices">
-
-    <jaxrs:serviceBeans>
-      <ref bean="helloService#default"/>
-    </jaxrs:serviceBeans>
-
-   </jaxrs:server>
-
-</beans>
