--- PsychMonkey7/PsychMonkey7a-api/src/main/java/edu/asu/bmi/cad/PsychMonkey7a/api/HelloService.java
@@ -1,38 +1,0 @@
-package edu.asu.bmi.cad.PsychMonkey7a.api;
-
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
-/**
- * HelloService!
- */
-@Path( "HelloService" )
-public interface HelloService
-{
-    @Path( "sayHello/{who}" )
-    @GET
-    @Produces( { MediaType.TEXT_PLAIN } )
-    String sayHello( @PathParam( "who" ) String who );
-}
